Day #1029: Consistency. Bugfixes/experiments.

+++ side_experiments/consistency_w_result/solution/obs_node.cpp
@@ -168,8 +168,8 @@ void ObsNode::clean() {
 }
 
 void ObsNode::measure_update() {
-	this->average_score = this->sum_score / this->num_measure;
-	this->average_instances_per_run = this->num_measure / MEASURE_ITERS;
+	this->average_score = this->sum_score / (double)this->num_measure;
+	this->average_instances_per_run = (double)this->num_measure / MEASURE_ITERS;
 }
 
 void ObsNode::save(ofstream& output_file) {
